fix: correct when we validate both client and server certs as part of TLS
Before we were incorrectly attempting to validate certificates that didn't exist. We were also ignoring validation of server certs that didn't exist.

We've had to split the server and client TLS config setup as the verifier needs to know which of server of client it represents to do the correct thing.