feat: allow for (and use) bcrypt for secret hashing
This change was way bigger than intended because our old abstractions for validating secrets assumed that given a plaintext secret you could deterministically generate the same hash each time. With bcrypt, and any other mechanism using salts, this is not true. To fix this we need to know the user/tenant/etc id in order to find the additional information that goes into a hash - or alternatively we need to move the validation of secrets behind the abstraction to allow salts to be embedded in the secret hash itself as bcrypt does.

This change makes these changes and removes the old abstractions