feat: allow independent configuration of token server and policy enforcement
Before if we weren't enforcing policies then we wouldn't bother to setup the token server - because why use tokens if we aren't going to validate them. This change separates these concerns allowing us to run a controller with the token server but not policy enforcement which should make testing easier.

I've also introduced separate config for local and tenant token support so we can turn these on/off as we wish