feat: allow manager connections to be saved by use sites without risk of using out of date connection data
Before we solved this problem using a func to get the manager address and calling this func each time we setup an RPC, each and every time. This meant we couldn't cache service clients or pass them through to other funcs or types - for example we couldn't pass `TenantApiClient` around, we had to pass `func() (*grpc.ClientConn, error)` instead.

With this change we replace the func with a node.Remote which has the ability to update the manager address in the internal state of grpc connections using the Resolver feature via node.DialChan. Now we can treat manager connections as lazy, optional, and auto-updating whenever enrollment happens or changes.