refactor: make use of externally defined enums & bitfields clearer
The DALI Bridge Go component interacts with a lot of data structures
from the DALI Bridge PLC component and the Tc3_DALI PLC library.
In order to do this correctly, enum definitions must match exactly
between the two systems. To make it clearer that this is the case,
and make it harder to break accidentally, avoid the use of `iota` for
these enums, instead listing all values explicitly. Document the
location of the reference documentation for the corresponding PLC
data structures.

Extract the occupancy input event bitfield extraction to a documented
helper method, to make the logic clearer.