dev(tc3dali): add context-aware event initialisation
Add a new abstraction to tc3dali: the 'once' type works like
sync.Once, but allows a failed process to be retried on the next
call and is context-aware so a call which is waiting can bail out
early.

Use the new 'once' type in controlDeviceServer for context-aware
initialisation of the DALI occupancy events. Previously, the
initialisation could block for up to 5 seconds.