dev(auto/testlight): new data structures for emergency light log
Data stored by testlight automation changed.

The automation stores the most recent status for each emergency light, keyed by the light's name.
This facilitates common queries like "is light X OK right now" and "when was the data for light X last updated".

There is also a log which stores events:
  - function test passes
  - duration test passes
  - status changes (only recorded when the latest status is different from the previous status to avoid cluttering the DB with useless data)

New structure allows runner.process to be more straightforward and easy to follow. Refactored the polling into pollDaliEmergency.

Added some utilities for dealing with slices of faults. Protobuf requires them to be slices, but semantically they are sets, so added utility functions for order- and count-independent comparisons, and normalisation.