refactor: make some util packages public
  - util/minibus
  - util/state
  - util/pki
  - util/jsontypes
  - util/rpcutil
  - db
  - manage/enrollment
  - manage/tenantapi

Some packages in bsp-ew depend on these, so the migration
cannot be completed until these are public.

This is a bit excessive, so some reorganisation is in order.
I hope to make more things private again before the open source
release, as it's harder to change public APIs before that.