feat: support acknowledgement authors
This updates the notification pages of the ui and the alert api to support recording and displaying who acknowledged an alert as per the designs.

Alert.Acknowledgement now accepts an Author which includes information about the actor who called the acknowledgeAlert api. The ui provides this information when calling the api from the claims of the logged in user. If an author is present the ui also displays this information in both an avatar and in a new popup menu that contains more information and the ability to unacknowledge the alert.