feat: add Node.Support(Api(...))
This changes how APIs are registered via the Node.Support api. Before the only way to get a new grpc api registered via a node was to implement it via a Router. Not all API support this pattern.

Now you can explicitly add support for a non-router api to the node.

The registration of routers as apis the node exports is handled for you via the Support(Router) API as before, though the logic is now separated into a addRouter and addApi call.