fix: avoid callers metadata from being mutated by Nodes via HasMetadata
This surfaced as metadata being mostly empty for devices configured via the mock driver.

What was happening was the mock driver, when configured to setup a device, would call Announce with the HasMetadata option using Device.Metadata. If this is the first time the device is seen the Node will create the relevant metadata client and add it to the metadata router, all via the Announce mechanism again. One of the steps of announcing a new device is to merge any metadata together to form a cohesive whole, this includes both the metadata that came from the mock drivers config, and the metadata the Node creates for you when it sets up the metadata client (i.e. to indicate the device implements the metadata trait). The issue is in the resource package, when setting values without a field mask values are set via reference without cloning (i.e. r.value = newValue), this causes the value stored in the nodes router for metadata to be the same object reference as the mock drivers Device.Metadata config! What happens next is that metadata is updated to reflect that it now supports the Metadata trait which gets added to both the node router and to Device.Metadata.

The next thing the mock driver does is go through all the configured traits (via Device.Metadata.Traits) and creates mock clients for each one. Unfortunately by this point that config also includes an unexpected record for setting up a mock Metadata client, which the mock driver does, and which blindly overwrites the metadata client setup and written to already in the router, causing all that info to be lost.