fix(ui-gen): make the js gen consistent with the go gen wrt the base dir
Before the go gen would use the ./proto dir as the base and js would use .
This meant there was no consistent way for one project proto to import another between the go and js generators.
I've made the js one match the go one as it simplifies things a bit in the generated code.