dev(ui/zones): Updates following DeviceTable changes + set zone devices
Some more work needs to be done to actually save the zone device list to the server - there are some
complications here in terms of how to know which 'group' a device should go in. It may be possible
to guess based on the supported traits, but this doesn't feel like a guaranteed approach.