fix(tls): disable go built in client cert verification by default, use our own
This fixes an issue, that I'm sure wasn't a thing, where our VerifyConnection call wasn't being called in favour of using the Go built in client cert verification logic. Logic that would use tls.Config ClientCAs which we don't populate as we get them during validation to support runtime changes to the certs.