fix(auto/lights): refresh lights immediately when changing modes
This fixes an undesirable behaviour where changing modes in an automation that's controlled with a button won't immediately apply the new mode. Instead, you have to switch the automation off and then on again for the new mode to apply.

Add parameter to NewWriteState that sets the initial button StateChangeTime cutoff. Use time.Now() in existing calls, as this maintains the same behaviour.

Cache most recently asserted mode in WriteState. This records the mode that was used when we last updated the brightnesses, allowing detection of a mode change.

Store the active mode in WriteState when performing an action.

Add logic to detect when the mode changes in decideAction. When the mode has changed and the lights were already on, refresh the light levels using the new mode.

Add test cast to demonstrate that this functions when using a button.