fix(node): explicitly check for the metadata router type, not just that it accepts metadata clients
The bug was surfaced when we added all the Info traits, some of which have no methods. This makes the routers have no methods which means their HoldsType methods return true for all client types, including those that do have methods. This caused the metadata logic to fail as the first empty router could be returned from metadataApiRouter