fix(node): race condition if more than one node.Announce is called with the same name breaking metadata (and hence Devices api)
If two drivers (or other systems) announce a name at the same time there was a race of the form `if !exists() then create()` that could cause more than one create call to be executed. This issue was exasperated because the new item made by create() was returned directly causing whatever edits were made to that model to be lost.

The symptom was that some devices would not announce that they implemented certain traits and we'd see duplicated "device now implements Metadata" in the logs.

I've added locking around the relevant places and attempted to avoid deadlocks by adding HasNoAutoMetadata feature to Node.Announce.