fix(zone): simplify the zone status logic significantly and fix race conditions and delays
This is at the cost of doing more api calls than needed, especially to devices that we know don't exist. I've adjusted the logs to hide this from daily observation but it's still there.

We should really keep track of the devices that consistently report that they don't implement status and don't ask them anymore. This is complicated so for now we'll just ignore that and