pki: play nicely with VerifiedChains during TLS
Use the correct API for loading certs for server side tls config instead of fudging it.

Before we were taking over the verification of certificates ourselves, but we don't have to. Doing it ourselves resulted in fields like TLSInfo.VerifiedChains from not being populated. Moving to the GetConfigForClient api means a) we aren't rolling our own server tls verification, and b) downstream code works as expected.