app: add --policy-mode=[on,off,check] flag
This flag supersedes the old --insecure-disable-policy flag which is now equivalent to --policy-mode=off.
The new --policy-mode=check performs authorization only if an identity (cert or token) is present, otherwise just lets the request through. This is insecure, but does allow gradual migration from insecure (off) to secure (on) without breaking everything.