opsui/login: updated authentication, navigation, redirect and error handling
Extended the local authentication functionality. On a successful login via local authentication, we are storing those values in localStorage, which is going to be returned in the Pinia state for the use of gRPC authentication for example.

Added `account.js` initialisation - setting up/making available all authentication providers. Depending on the result of this initialisation, we then set all available auth provider types in the `availableAuthProviders` ref, this way we can hide/show them on the login page.
However, the initialisation process can cause a delay in displaying the available options on the login screen (if waiting for the promise to be resolved but the server/auth provider not available).
If this happens, an error message (snackbar) will pop up and the list of provider options are going to exclude the unavailable ones.

Finally, the `grpcweb.js` got a permission/unauthenticated switch with cases, so we can handle errors better during navigation and/or on api requests.

Jira: SC-79, SC-80