opsui/auth: ensure init has been called before relying on state setup during init
This fixes a race condition where api calls (that use refreshToken) can happen while the auth providers are being initialised, what this means is those API calls can fail with "NOT_AUTHENTICATED" because we haven't read the token yet. This flow ultimately results in the logout flow being invoked which can cause you to log out on page load, even when you don't need to.

In the most confusing cases this presents as the redirect logic that is supposed to send you back to the page you were on after login not working.