opsui/overview: use `EnergyGraphOptionsMenu`
Add the newly created `EnergyGraphOptionsMenu` to the line chart and add functionality, so we can change the duration of the data we display (24H / 1W / 30D).
Also, removed `span` and `timeFrame` props because they are not used anymore. Instead, we are hardcoding the available duration options - which will provide the `span/timeFrame` values we use for data fetching - replacing these removed props.

Jira: PRJ-117