opsui/overview: add durations to `useCarbonIntensity`
After updating the line chart with the ability to select different duration, this PR is going to connect the duration change with the carbon data fetch.
We pre-fetching the carbon intensity values on page load for all 3 duration types, then we calculate - depending on the active duration's id - which carbon intensity data we need, and if the conversion is toggled - we use the matching one.

Jira: PRJ-117