opsui/overview: add click handler to menu option list item
As per requested on review, the option list items are handling the click events, rather than the switch toggle/button group.
Also added debounce to the data fetching logic, this way we can wait for the user to settle down on an option / reduce the number of fetch calls if the user quickly swaps between duration option.

Jira: PRJ-117