opsui: remove console error statements for cases where we expect errors
Checking for hub nodes or checking enrollment are both operations that we do optimistically. When they fail this tells us that the server isn't a hub or node, this is not an error condition. When we see hundreds of error messages in the dev tools console it's useful to know that they are actually errors, so don't print these cases anymore.