opsui/energy: only use properties we need in script setup
When using Vue DevTools, the tools will take your component and stringify the entire state, transfer that state to the dev tools process, then unstringify it. What this means is the data contained in your component state directly impacts the performance of dev tools; and as a consequence you application too.

In the EnergyGraph component we had two copies of the useMeterHistory composable exported as data in the component. These composables contain both the full record set retrieved from the server and a processes "series" data set to display; the component only needs the series data but exported it all. Because of this, when dev tools was open and looking at the component, the entire record set (1000s of objects) would all be stringified and transferred to the dev tools on each tick causing ~300% CPU and GBs RAM to be used. This change makes sure we only export the series data drastically reducing both memory and CPU usage via dev tools for this component.