opsui/ui-config: refactor (example) ui-config and ui-config store
Refactoring the example ui-config/building object to follow a certain logic. This involves moving the existing `building` object into the `ops` object, and renaming it to `overview` - as it is part of the `ops/overview/building` path and its sub-paths. It also includes renaming the `traits` to `widgets`, alphabetical reordering and setting default brandColors.

Additionally, a change required because each project and building can have different setup. Not all building has 'headcounting' capability (occupancy bar chart) or generated power etc.
Following the structure of the overview and its children (floors/zones), we can show/hide the widgets we want to display on the building overview page.

Jira: N/A