opsui: fix legend rendering during some navigations
There appears to be some inconsistency in how chart.js plugins expose their data for legends and how that interacts with Vue reactivity. After some investigations the old computed property that was making the legend items (the data that is rendered) was being called twice after each plugin.afterUpdate call. The first would work but by the time the second was called the internal state of the chart had changed and was inconsistent: the legend thought it had two datasets but the data only had one. This caused the exception and broke the render.

I've moved the legend data computation into the plugin hook and records the data in a ref instead of computed property which seems to have fixed the issue.

Issue: SC-512