opsui/ops: allow complete configuration of the ops overview section
This change adds a new config mechanism for the ops/overview pages. Previously there was some configuration available, but a lot of it was hard coded - especially the main building overview page. Config of each page was quite limited and generally relied on variations on names to fill out all the devices that each widget on the page would need.

The new config is now explicit about both the layout and the widget configuration. All pages under /ops/overview are now configurable with no distinction between the top level and child pages. Layouts can now be chosen and are extendable and each widget is fully configurable via json and vue props. Any hard coded config is now simply a default and can be provided via the ui config. New widgets are easy to add and are loaded dynamically by the ui.

The old config is still supported (for now) but the ui converts it into the new format when used. If an ops.pages config property exists it is assumed that the new config approach is used and any legacy properties (buildingZone, overview, etc) will be ignored.

I've added an example ui config that matches the vanti-ugs app config.

Issue: SC-510
Issue: SC-508