opsui/alerts: fix the notification table re-rendering on alertMetadata update
There's a reactivity chain between updates to the alert metadata and the query being run to fill the table, via the query filter component which needs to know the list of floors, zones, etc.

The alert metadata updates quite frequently, usually just a count here or there but the filter doesn't check that and recalculates the current query which re-renders the table.