opsui: fix dragging the sidebar was janky
This was caused by the logic using incorrect apis to perform the drag. It was attempting to use document events to drag things, instead it should have been using el.setPointerCapture. This API is designed for this type of thing and even works if you drag outside the browser window. No more issues and things feel much more responsive.