opsui/dash: support arbitrary queries when using the notification widget
Previously only zones were supported as a configuration mechanism - zone is still the only option when using the old dashboard config structure. Now you can specify a "forceQuery" parameter that sets the underlying query to use for the table rows and can include anything from the query proto object.

This allows us to show all issue on a floor, for example.

Issue: SC-511