opsui/traits: replace cleanup onUnmounted with onScopeDispose
The latter allows these composables to be used outside a Vue component, for example from within a watcher callback.

We also now don't pretend that we don't have a resource leak in the latter case as the watch call was not being unwatched when not in the correct context, though you'd get a warning before about `onUnmounted not available outside script setup` - you get a better warning now.