opsui/sidebar: move sidebar.sidebarNode to services store
This data was never really related to the sidebar, it was used as a way to communicate which hub node we were displaying. It just happened that the config editing functionality in the sidebar needed the info too.

Moving it to services has the side effect of remembering which node we have selected between automations, drivers, and services (and zones) though I'd like to remove this distinction at some point in the future.