opsui/sidebar: fix that data could leak when clicking on notifications
Before the existing sidebar data was included when clicking on a notification, the intent being to save metadata/notification data as the other was fetched asynchronously. This leaked any existing sidebar data added by other components. The solution is to only keep the data we know about in the notification-sidebar comms instead of using ...spread.