actions: fix multiarch container builds
Multiarch containers are only uploaded correctly
by docker/build-push-action if they are all
done together.

If uploaded in multiple steps, the newest will
overwrite the oldest and the image is not multiarch.

Tweak Dockerfile to look in arch-specific directory
for copying in build artifacts. Modify action
to extract to the correct directories.