conductor: support passing in GIT_VERSION from environment
If building in an environment where the Git history is not available,
the build previously failed because the version could not be determined.

Support using environment variable GIT_VERSION to set/override the
version number injected into the build.

Remove dependency 'git-revision-vite-plugin' and implement functionality inside vite.config.js.
The external plugin is obscure and documented entirely in Chinese,
so I couldn't find a way to make it work in the absence of a Git
repo.