docker: add a Dockerfile that can perform the entire build
Unlike prebuilt.Dockerfile, which copies prebuilt artefacts into the
container, this Dockerfile performs the entire build internally.
This is useful where the user doesn't have the full development
environment available, but still wants to perform a build.

It requires NPM credentials to fetch the dependencies.
These are injected by using mounting .npmrc as a secret.