auth: use protojson to marshal requests into policy inputs
To correctly convert proto messages to OPA policy data we need to reliably be able to convert those messages to json.
The protojson package is a much better fit for this than the json package, which is the default used by OPA.
Unfortunately there's no easy way to hook into specific type marshalling of inputs, so we convert it before passing to the OPA evaluation.