system/proxy: improve logging for cancelled sub-tasks
Before when a task spawned sub-tasks but could also fail itself, for example announceHub, those subtasks would report as WARN: task failed, will retry.
This was inaccurate as the error reported would be "context cancelled" and the task would not be retried.

The new code doesn't log, though if we want to enable that in the future we can do.