test: fix race conditions in tests
Some tests fail to pass under the race detector.

Several of them were working before due to a deliberate pause between results being written and being checked.
However, this isn't technically allowed by the memory model.
Fix the errors by introducing synchronisation.

serviceapi: fix send on closed channel

This is a subtle one: Api.pullServices can produce a send on a closed channel.
This doesn't seem to occur in real-world use, and was detected using the race detector.

The out channel is closed in the main goroutine. Both the main goroutine and extra watchRecord goroutines send on the channel.
The code attempts to guard against a send on closed channel by using contexts to stop the watchRecord goroutines at the same time close(out) is called.
However, this is invalid because out still appears in the select statement, even if the <-ctx.Done() case is ready.

Introduce another channel outSem to guard access to the out channel.
Using this channel, we ensure that a watchRecord goroutine can never get access to the out channel after it's closed.

service: slacken test timeouts

Enabling the race detector has caused test timeouts to trip, as the race detector slows execution quite a lot.

Increase tt.waitForState timeouts to avoid running in to this issue.

lights: fix race condition in test

Use a mutex to protect access to testActions.calls/nextCall.
The race detector was tripping on these accesses on Go 1.20 (but not Go 1.22).