node: auto register Parent and Metadata apis when required
Previously it was the responsibility of the caller to ensure that the node `Support`ed the ParentApi and MetadataApi before calling Announce.
This was prone to error and forgetfulness, even if it was easy to fix.

This change makes those apis (and their Info equivalents) automatically registered with a grpc server if not already registered by the caller.