task: stop polling immediately when context is cancelled
Because select is nondeterministic when multiple cases are ready,
Poll/PollErr might run one time extra after the context is cancelled.
This is because the ticker channel may already contain an element sent during the previous poll operation.
Add an explicit check for a completed context before checking for a tick.
This reduces the chance of an extra run (it is still possible if the cancellation happens in the window between the select and the action run).

Allow 1 extra run in TestPollErr/stops

Fixes SC-606