block: fix error patching field of array element
If a patch references a sub-field (or sub-array, at any nesting depth) of an array element that doesn't exist, the patch failed to apply.

In applyPatch, when referencing an array element that doesn't exist, create a placeholder with the right key so we have something to apply the patch to.

This could be triggered with patches from Diff, if the patch to the sub-field came before the patch creating the array element.