appconf: prevent Store corruption in BootConfig
There is a corruption condition that could occur in a Store if a method returns an error half way through BootConfig.

Assuming the Store contains both a local and active config:

1. BootConfig swaps in newLocal as the Store's local config (saved on disk)
2. Store.SetActiveConfig(...) fails or is interrupted
3. BootConfig called again - newLocal is now identical to saved local config in the Store
4. No diffs are applied to active config
5. Active and local config are now desynchronised.

Split Store.SwapLocalConfig into GetLocalConfig and SetLocalConfig.
Re-order operations so that local config is only written once all other operations have succeeded.
If store.SetLocalConfig fails, then the same diffs will be applied again to the active config on the next call to BootConfig.
The block patches are idempotent, so applying them for a second time does no harm.

Add test to block that checks block.ApplyPatches is idempotent.