appconf: improve config file update edge cases
Atomically replacing a file is a subtle thing to do.
Use renameio to implement this, which correctly handles:

  - Setting permissions
  - Removing the temporary file on failure
  - Flushing writes before renaming the file

It's not possible to atomically replace a file on Windows
(https://github.com/golang/go/issues/22397#issuecomment-498856679)
but we only need the atomicity guarantees on our production platforms
(Linux and FreeBSD).

renameio is a simple and stable library.
I have reviewed the relevant code path and it handles edge
cases better than the trivial writeFileAtomic function.