system/gateway: rename proxy system to gateway
We call a node that uses the proxy system a gateway.
Rename the system to 'gateway' to make this clearer,
and to differentiate it from the proxy driver.

The default system set still recognises 'proxy' as a valid name
for the gateway system, to preserve compatibility with existing
system config files.