ui/ops: fix map lookup issue in useCollection
Before the getIndex function was returning undefined instead of -1 when an id wasn't already known. This would cause all items to be put into the undefined index which had terrible performance and didn't work. It only appeared to work because the pull would eventually get all the items anyway.

This fix changes the pull to only pull updates and fix the map lookup to return -1 if not found. Now things work