node: implement router-based gRPC service handling
Refactor metadata implementation to use allMetadata as
the source for the server.

Add Node.ClientConn(), the new preferred way for services
to get a client that routes using the node.

Remove node.{Api,Routing,Clients} features.
Add new RoutedService and UnroutedService featutes.
Unlike the old features, these add support for routing without
adding an actual server implementation.

The new features are often not required because support for
a service will be added automatically when the service is
announced on a node.

Adapt services to use the new API.