direct trait requests without a name to the node
name.IsAbsent{Unary,Stream}Interceptor
no longer works because the routing
interferes with receiving the message.
It is first decoded as an emptypb.Empty,
so the name property is unavailable.

Reimplement as part of the node & router,
by mapping an empty name to the node
name before processing the request.