node: when announcing a device with a trait, support all that trait's services
When a device has a trait, make the router support all
the services associated with that trait, even if the
announced device doesn't have servers for all of those
services.

For example, even if the announcement only has a server
for the XxxApi service, also support the XxxInfo service
on the router, so it will be visible over gRPC reflection.