router: use correct descriptor when copying messages
Do not use emptypb.Empty as an intermediate message when
copying between client and server streams for
StreamHandler.
The problem is that fields cannot be looked up by name
in interceptors. This breaks our policy interceptor
because it needs to pass a JSON representation of the
request message to the policy engine.

Instead of emptypb, use dynamicpc.
In Method, use protoreflect.MethodDescriptor instead
of StreamDesc, so that the request and response
MessageDescriptors are available.