app: provide correct stream configuration to interceptors
The policy interceptor needs to know what kind of
stream configuration an RPC is using to do the correct
kind of policy checks.

Streams handled by the UnknownServiceHandler always
report as bidirectional streams.
The router knows what the actual stream configuration
of all its services are; add StreamServerInfo method
to fetch it.

interceptors: add CorrectStreamInfo interceptor to
look up and apply the right stream configuration

app: use CorrectStreamInfo interceptor as the first in
the chain