driver/proxy: don't advertise 'skipChild' devices
Any device that has metadata will be visible via the Parent and Devices API. As these apis are used for proxy and gateway functionality this can cause issues where the owner node for a device isn't the one you'd expect, something that depends on boot order and restarts.

This change makes sure no metadata is added for proxy nodes configured with skipChild.

Part of the fix for SC-909 and SC-908