devices: allow downloading a CSV of devices and their trait data
This change introduces a new DevicesApi download function that can be used to retrieve a CSV file of a filtered set of devices (like you get on the devices table) along with selected information gathered from their traits.

The devices table now has a new Download CSV button that will use the current filters (floor, subsystem, etc) as the filter for the device data to download.

The implementation uses two steps:
1. Use gRPC to retrieve a URL where the file can be downloaded from. The gRPC request accepts the various options for what data should be downloaded, like the device query.
2. GET that URL to download the CSV file.

The reason for the two-step process is to allow the browser to stream the CSV download, avoiding the need to collect all the information into memory before constructing the CSV.

While the gRPC step is authenticated, the GET URL step isn't. We need to think of a way to reduce access to this link, either via time or securing the access directly. I suspect adding an expiry and signature to the download url will solve this problem reasonably well.