node: add AnnounceScope and ReplaceAnnouncer to allow properly sequencing announcements with undos
AnnounceContext has a flaw which makes it impossible to sequence announcements after undos. Announcements are undone asynchronously when the context completes, which can conflict with new announcements to the same name.

Add AnnounceScope to handle the common case of a group of announcements which can all be undone at once, synchronously.

Add ReplaceAnnouncer to help the case where a service wants to re-register all its announcements when it restarts, without conflicting with the undoing of previous announcements.
It is also context aware so announcements can be automatically undone when the service terminates.