policy: add HTTP middleware to authorize access
Add HTTP support to policy.Interceptor.

Add Protocol and Path to Attributes to allow policies to access HTTP details.

Only hook up HTTP auth for the log level & version (which is always allowed) handlers for now.
Other endpoints should have it applied, but we need to do a comprehensive inventory of endpoints first and work out what to do with the dynamically registered endpoints.