auto/lights: attempt to remove inaccurate "pulls are now succeeding" logs
Our logs end up being filled with:
> pulls are failing...
> pulls are now succeeding...
> pulls are failing...

This is caused (I believe) by non-uniform failure times. As the pull itself is blocking on success we have to set up a watchdog that expires after some time to know if the pull has blocked, instead of returning an error. The time the watchdog waits is variable, we have no idea how long failure takes for the pull function.

To mitigate this we time how long a failure takes to happen and configure the watchdog to wait longer than that until it thinks the pull is successfully blocking. I think it's this wait that is incorrect, so add logging and increase the timeout to try and fix the incorrect "success" logs.