auto/lights: rename BrightnessWriteState to Value
This makes the value more reusable for any other states we might want to record. It also makes the lighting automation match the patterns in the newer bms automation.

Note: this also changes the logic slightly to record write failures in the write state. We aren't making use of this change yet, but will do when we start logging more detail less frequently.