auto/lights: better logging
In summary, collect all log messages into the write state, then decide whether to log them all in one go or not.

In more detail, add the concept of reasons to write state and the process trigger so we can capture why the logic made the decisions it did and why it was running in the first place. The concept was first written in the bms auto.

Secondly add a new actions implementation that counts and collects information on the action calls so we can later ask "has anything changed" before logging. I've also added a few other utility actions implementations that help with logging verbose writes, and testing without affecting the real devices.

Finally update the config struct so that we can adjust the logging behaviour to meet our troubleshooting needs.