auto/lights: add non-state based trigger reasons to the write log
If the state is being processed due to reasons excluding state changes, log that fact in the output log. We may decide in the future to also log state changes in the completion log, but for now it's quite involved to calculate and can be large depending on how many things the automation is monitoring.