manage/devices: be more explicit about errors in parsePath
I figure it's generally a good idea to be explicit about what error is being returned and test for this.

It can also be a bad idea to return `errors.New("some error")`

According to the documentation:
>Each call to New returns a distinct error value even if the text is identical
Example where this could be an issue: https://go.dev/play/p/MDAsU493G7n.