system/gateway: use local descriptors where possible
Change the ServiceDescriptor used by the gateway to the one in protoregistry.GlobalFiles if it exists. This makes our dynamic services more consistent with assumptions in the proto package (proto.Equal, proto.Merge) which assume a message is described by a single service descriptor, this wasn't true before this change.

I've implemented the change in a way that avoids additional reflection calls on the remote node if we already know how to describe the service.

Fixes SC-1004